Move interpolation-related utils out of States into ale.h and update tests appropriately
